fix: changed get_roles to Campaign Handler, changede migration file
